@@ -5,20 +5,18 @@
   "license": "GPL-3.0-or-later",
   "dependencies": {
     "i18next": "^22.0.6",
-    "i18next-browser-languagedetector": "^7.0.1",
     "leaflet": "^1.5.1",
     "leaflet.locatecontrol": "^0.72.0",
     "leaflet.markercluster": "^1.4.1",
-    "prop-types": "^15.7.2",
-    "react": "^16.8.6",
-    "react-dom": "^16.8.6",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-helmet": "^6.1.0",
-    "react-i18next": "^12.0.0",
+    "react-i18next": "^12.1.1",
     "react-leaflet": "^2.4.0",
     "react-leaflet-markercluster": "^2.0.0-rc3",
     "react-markdown": "^8.0.4",
     "react-router-dom": "^5.1.0",
-    "react-scripts": "3.0.1",
+    "react-scripts": "5.0.1",
     "rehype-raw": "^6.1.1"
   },
   "scripts": {
@@ -26,7 +24,8 @@
     "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject",
-    "predeploy": "npm run build"
+    "predeploy": "npm run build",
+    "lint": "eslint src --ext .js --fix"
   },
   "eslintConfig": {
     "extends": "react-app"
