@@ -1,4 +1,4 @@
-import React, { Component } from 'react';
+import { Component } from 'react';
 import ErrorScreen from '../components/ErrorScreen';
 
 class ErrorBoundary extends Component {
