@@ -1,49 +1,18 @@
-import React from 'react';
+import { useContext, useState, useCallback } from 'react';
 import styles from '../css/search.module.css';
 
 import { MapContext, HotelContext } from '../context';
 import { useTranslation } from 'react-i18next';
-/**
- * @param {{name: string, address: string, mainOligarch: {name: string}[]}} place
- * @param {string} phrase
- * @returns {*|boolean}
- */
-function findProperty(place, phrase) {
-    let foundOligarch = [];
-    if (place.mainOligarch.length > 0) {
-      foundOligarch = place.mainOligarch.filter(oligarch => {
-        return (oligarch.name.toLowerCase().includes(phrase) ||
-          removeAccents(oligarch.name).includes(phrase));
-      });
-    }
-    const foundPlace = place.name &&
-      (place.name.toLowerCase().includes(phrase) ||
-      removeAccents(place.name).includes(phrase));
-    const foundAddress = place.address &&
-      (place.address.toLowerCase().includes(phrase) ||
-      removeAccents(place.address).includes(phrase));
 
-
-    return (foundPlace || foundAddress || foundOligarch.length > 0);
-  }
-
-/**
- * Source: https://stackoverflow.com/questions/990904/remove-accents-diacritics-in-a-string-in-javascript#answer-37511463
- *
- * @param {string} string
- * @returns {string}
- */
-function removeAccents(string) {
-  return string.toLowerCase().normalize("NFD").replace(/[\u0300-\u036f]/g, "");
-}
+import findProperty from '../utils/search/find-property';
 
 function Search() {
-  const {dispatch} = React.useContext(MapContext);
+  const {dispatch} = useContext(MapContext);
   const { t } = useTranslation();
-  const {hotels} = React.useContext(HotelContext);
-  const [value, setValue] = React.useState('');
+  const {hotels} = useContext(HotelContext);
+  const [value, setValue] = useState('');
 
-  const onSearchCallback = React.useCallback((event) => {
+  const onSearchCallback = useCallback((event) => {
     event.preventDefault();
     dispatch({ type: 'TogglePopup', showPopup: false });
 
@@ -52,7 +21,7 @@ function Search() {
     dispatch({ type: 'ToggleList', showList: true });
   }, [dispatch, hotels, value]);
 
-  const onKeyUpCallback = React.useCallback((event) => {
+  const onKeyUpCallback = useCallback((event) => {
     setValue(event.target.value);
     if (event.key === 'Escape' || value === '') {
       dispatch({ type: 'SetList', list: [] });
