@@ -1,13 +1,13 @@
-import React from 'react';
+import { useEffect, useContext } from 'react';
 import Header from '../components/Header.js';
 import List from '../components/List.js';
 import Menu from '../components/Menu';
 import { MapContext } from '../context';
 
 const Layout = (props) => {
-  const {dispatch, showList } = React.useContext(MapContext);
+  const {dispatch, showList } = useContext(MapContext);
 
-  React.useEffect(() => {
+  useEffect(() => {
     dispatch({ type: 'ToggleMenu', showMenu: false });
   }, [dispatch]);
 
